Server/requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart  # for file uploads

# LangChain & Ecosystem
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface

# Vectorstore
pinecone

# Embeddings via Hugging Face API
huggingface_hub

# PDF Parsing
PyPDF  # backend for PyPDFLoader via langchain_community

# Environment Variables
python-dotenv

# Typing & Utilities
pydantic
requests
tqdm

# Authentication
pymongo[bson] 
bcrypt
